requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv